Fix Message2Call constructor and unit-test assertions.
Keep the send lambda last so existing trailing-lambda call sites compile, and compare OkHttp pingInterval as Long.

+++ core/data/src/test/kotlin/io/github/nutea/anylisten/core/data/NetworkFactoryTest.kt
@@ -14,7 +14,7 @@ class NetworkFactoryTest {
     @Test
     fun clientPingsWebSocketsSoHalfOpenLinksFail() {
         val client = NetworkFactory.client()
-        assertEquals(NetworkFactory.WEBSOCKET_PING_INTERVAL_MS, client.pingIntervalMillis)
+        assertEquals(NetworkFactory.WEBSOCKET_PING_INTERVAL_MS, client.pingIntervalMillis.toLong())
         assertTrue(client.retryOnConnectionFailure)
     }
